+++ inc/commoninjectionlib.class.php
@@ -27,6 +27,8 @@
  * @link      https://github.com/pluginsGLPI/datainjection
  * -------------------------------------------------------------------------
  */
+
+use Glpi\DBAL\QuerySubQuery;
 use Glpi\Exception\Http\HttpException;
 use Glpi\Features\AssignableItem;
 
@@ -1865,7 +1867,6 @@ class PluginDatainjectionCommonInjectionLib
         /** @var DBmysql $DB */
         global $DB;
 
-        $where    = "";
         $continue = true;
 
         $injectionClass->getOptions($this->primary_type);
@@ -1893,17 +1894,15 @@ class PluginDatainjectionCommonInjectionLib
             if (!$continue) {
                 $this->values[$itemtype]['id'] = self::ITEM_NOT_FOUND;
             } else {
-                $sql = "SELECT *
-                    FROM `" . $injectionClass->getTable() . "`";
-
                 if (!is_a($itemtype, CommonDBTM::class, true)) {
                     throw new HttpException(500, 'Class ' . $itemtype . ' is not a valid class');
                 }
-                $item = new $itemtype();
+                $item  = new $itemtype();
+                $where = [];
+
                 //If it's a computer device
                 if ($item instanceof CommonDevice) {
-                    $sql .= " WHERE `designation` = '" .
-                    $this->getValueByItemtypeAndName($itemtype, 'designation') . "'";
+                    $where['designation'] = $this->getValueByItemtypeAndName($itemtype, 'designation');
                 } elseif ($item instanceof CommonDBRelation) {
                     //Type is a relation : check it this relation still exists
                     //Define the side of the relation to use
@@ -1919,49 +1918,41 @@ class PluginDatainjectionCommonInjectionLib
                         $source_itemtype      = $item::$itemtype_1;
                         $destination_itemtype = $item::$itemtype_2;
                     }
-                    $where .= " AND `$source_id`='" .
-                    $this->getValueByItemtypeAndName($itemtype, $source_id) . "'";
+                    $where[$source_id]      = $this->getValueByItemtypeAndName($itemtype, $source_id);
+                    $where[$destination_id] = $this->getValueByItemtypeAndName($itemtype, $destination_id);
                     if ($item->isField('itemtype')) {
-                        $where .= " AND `$source_itemtype`='" .
-                        $this->getValueByItemtypeAndName($itemtype, $source_itemtype) . "'";
+                        $where[$source_itemtype] = $this->getValueByItemtypeAndName($itemtype, $source_itemtype);
                     }
-                    $where .= " AND `" . $destination_id . "`='" .
-                    $this->getValueByItemtypeAndName($itemtype, $destination_id) . "'";
-                    $sql   .= " WHERE 1 " . $where;
                 } else {
                     //Type is not a relation
 
                     //Type can be deleted
                     if ($injectionClass->maybeDeleted()) {
-                        $where .= " AND `is_deleted` = '0' ";
+                        $where['is_deleted'] = 0;
                     }
 
                     //Type can be a template
                     if ($injectionClass->maybeTemplate()) {
-                        $where .= " AND `is_template` = '0' ";
+                        $where['is_template'] = 0;
                     }
 
                     //Type can be assigned to an entity
                     if ($injectionClass->isEntityAssign()) {
                         //Type can be recursive
                         if ($injectionClass->maybeRecursive()) {
-                            $where_entity = getEntitiesRestrictRequest(
-                                " AND",
-                                $injectionClass->getTable(),
-                                "entities_id",
-                                $this->getValueByItemtypeAndName(
-                                    $itemtype,
+                            $where = array_merge(
+                                $where,
+                                getEntitiesRestrictCriteria(
+                                    $injectionClass->getTable(),
                                     'entities_id',
+                                    $this->getValueByItemtypeAndName($itemtype, 'entities_id'),
+                                    true,
                                 ),
-                                true,
                             );
                         } else {
                             //Type cannot be recursive
-                            $where_entity = " AND `entities_id` = '" .
-                            $this->getValueByItemtypeAndName($itemtype, 'entities_id') . "'";
+                            $where['entities_id'] = $this->getValueByItemtypeAndName($itemtype, 'entities_id');
                         }
-                    } else { //If no entity assignment for this itemtype
-                        $where_entity = "";
                     }
 
                     //Add mandatory fields to the query only if it's the primary_type to be injected
@@ -1969,44 +1960,59 @@ class PluginDatainjectionCommonInjectionLib
                         foreach ($this->mandatory_fields[$itemtype] as $field => $is_mandatory) {
                             if ($is_mandatory) {
                                 if ($item instanceof User && $field == "useremails_id") {
-                                    $email = $DB->escape($this->getValueByItemtypeAndName($itemtype, $field));
-                                    $where .= " AND `id` IN (SELECT `users_id` FROM glpi_useremails WHERE `email` = '$email') ";
+                                    $where['id'] = new QuerySubQuery([
+                                        'SELECT' => 'users_id',
+                                        'FROM'   => 'glpi_useremails',
+                                        'WHERE'  => ['email' => $this->getValueByItemtypeAndName($itemtype, $field)],
+                                    ]);
                                 } else {
-                                    $where .= " AND `" . $field . "`='" . (string) $this->getValueByItemtypeAndName($itemtype, $field) . "'";
+                                    $where[$field] = $this->getValueByItemtypeAndName($itemtype, $field);
                                 }
                             }
                         }
                     } else {
                         //Table contains an itemtype field
                         if ($injectionClass->isField('itemtype')) {
-                            $where .= " AND `itemtype` = '" . $this->getValueByItemtypeAndName(
-                                $itemtype,
-                                'itemtype',
-                            ) . "'";
+                            $where['itemtype'] = $this->getValueByItemtypeAndName($itemtype, 'itemtype');
                         }
 
                         //Table contains an items_id field
                         if ($injectionClass->isField('items_id')) {
-                            $where .= " AND `items_id` = '" . $this->getValueByItemtypeAndName(
-                                $itemtype,
-                                'items_id',
-                            ) . "'";
+                            $where['items_id'] = $this->getValueByItemtypeAndName($itemtype, 'items_id');
                         }
                     }
 
                     //Add additional parameters specific to this itemtype (or function checkPresent exists)
                     if (method_exists($injectionClass, 'checkPresent')) {
-                        $where .= $injectionClass->checkPresent($this->values, $options);
+                        $extra = $injectionClass->checkPresent($this->values, $options);
+                        if (is_array($extra)) {
+                            if (count($extra) > 0) {
+                                $where = array_merge($where, $extra);
+                            }
+                        } elseif (!empty($extra)) {
+                            trigger_error(
+                                sprintf(
+                                    '%s::checkPresent() must return an array, %s returned instead.',
+                                    get_class($injectionClass),
+                                    gettype($extra),
+                                ),
+                                E_USER_WARNING,
+                            );
+                        }
                     }
-                    $sql .= " WHERE 1 " . $where_entity . " " . $where;
                 }
-                $result = $DB->doQuery($sql);
-                if ($DB->numrows($result) > 0) {
-                    $db_fields = $DB->fetchAssoc($result);
+
+                $result = $DB->request([
+                    'FROM'  => $injectionClass->getTable(),
+                    'WHERE' => $where,
+                ]);
+
+                if (count($result) > 0) {
+                    $db_fields = $result->current();
                     foreach ($db_fields as $key => $value) {
                         $this->setValueForItemtype($itemtype, $key, $value, true);
                     }
-                    $this->setValueForItemtype($itemtype, 'id', $DB->result($result, 0, 'id'));
+                    $this->setValueForItemtype($itemtype, 'id', $db_fields['id']);
                 } else {
                     $this->setValueForItemtype($itemtype, 'id', self::ITEM_NOT_FOUND);
                 }